refactor(gateways): resolve audit findings on the consolidation
Apply structural fixes flagged by kural audit on the refactor:

- Tag both adapter adaptPrice functions with @kuralPatterns
  gatewayPricingAdapter — they are template repetitions of one
  concept across gateways, not coincidental duplicates.
- Tag registry.ts with @kuralBound inward — its identity is dominated
  by the adapters it wires, which is a permanent architectural
  property, not an accident of vocabulary.
- Move CatalogFetchParams back to registry.ts. It was inadvertently
  hoisted into http.ts during the consolidation, but http.ts owns
  output shapes (price, throughput, catalog entry) while
  CatalogFetchParams is the input shape on the GatewayAdapter
  contract. Restoring it keeps ollama (which doesn't use the
  OpenAI-style flow) from coupling to catalogFetch.ts.
