+++ src/llms/gateways/vercel.ts
@@ -7,8 +7,8 @@
  * JSON field names or its catalog URL conventions.
  */
 
-import type { CatalogEntry, CatalogFetchParams, PricePerMillionTokens } from "./http.ts";
-import type { GatewayAdapter } from "./registry.ts";
+import type { CatalogEntry, PricePerMillionTokens } from "./http.ts";
+import type { CatalogFetchParams, GatewayAdapter } from "./registry.ts";
 import type { OpenAIStyleCatalogSpec } from "./catalogFetch.ts";
 import { fetchOpenAIStyleCatalog } from "./catalogFetch.ts";
 import { parsePerToken } from "./http.ts";
@@ -28,6 +28,7 @@ const REASONING_TAG = "reasoning";
  * @returns Per-million USD rates
  * @kuralPure
  * @kuralHelper
+ * @kuralPatterns gatewayPricingAdapter
  */
 function adaptPrice(pricing: Record<string, unknown>): PricePerMillionTokens {
   return {
